agentcodecraft/requirements.txt
fastapi==0.114.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.31
pydantic==2.8.2
pydantic-settings==2.4.0
pyyaml==6.0.2
streamlit==1.37.1
httpx==0.27.0
pytest==8.3.2
requests==2.32.3
google-genai==1.51.0
google-adk>=1.19.0
